# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API Documentation",
        "description": "Dokumentasi API e-commerce MK3",
        "contact": {
            "email": "kamu@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Ambil daftar transaksi customer yang sedang login",
                "operationId": "7e07cc202ed677881195e1df63ca2991",
                "responses": {
                    "200": {
                        "description": "Daftar transaksi berhasil diambil"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Buat transaksi baru",
                "description": "Membuat transaksi baru oleh customer, mengurangi stok produk dan menyimpan detail transaksi",
                "operationId": "a6ed73bc29cdce8903f4822f3e2f6a24",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "address",
                                    "payment_method",
                                    "products"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Jl. Merdeka No. 1"
                                    },
                                    "payment_method": {
                                        "type": "string",
                                        "enum": [
                                            "cash",
                                            "credit_card",
                                            "bank_transfer"
                                        ],
                                        "example": "cash"
                                    },
                                    "products": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 101
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaksi berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Transaksi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/transactions/{id}": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Ambil detail transaksi berdasarkan ID",
                "operationId": "7ffb0c851e0a20910cc9a97155724f21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail transaksi"
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/transactions/{id}/status": {
            "patch": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Update status semua detail transaksi",
                "operationId": "12d8e93d198d062980cc082133781b71",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "paid",
                                            "cancelled"
                                        ],
                                        "example": "paid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status transaksi berhasil diperbarui"
                    },
                    "422": {
                        "description": "Validasi status gagal"
                    },
                    "404": {
                        "description": "Transaksi tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Bearer Token via Sanctum",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}